Tab is now a module, better API and logging conventions
