+++ src/Chrome.ts
@@ -1,16 +1,11 @@
-import * as fs from 'fs';
 import * as _ from 'lodash';
 import * as chromeLauncher from 'chrome-launcher';
 import * as CDP from 'chrome-remote-interface';
-import * as path from 'path';
 import * as debug from 'debug';
-import { EventEmitter } from 'events';
 
-const events = {
-  done: 'done',
-};
+import { ChromeTab, events } from './ChromeTab';
 
-const log = debug('chrome');
+const log = debug('navalia:chrome');
 
 export interface chromeOptions {
   [propName: string]: boolean | undefined;
@@ -18,170 +13,6 @@ export interface chromeOptions {
 
 export interface customOptions {
   maxActiveTabs?: number;
-}
-
-export interface navigateOpts {
-  onload?: boolean
-}
-
-export type triggerEvents =
-  'click' |
-  'mousedown' |
-  'mouseup' |
-  'mouseover' |
-  'touchstart' |
-  'touchend' |
-  'focus' |
-  'touchcancel' |
-  'touchmove' |
-  'change' |
-  'blur' |
-  'select'
-;
-
-export class ChromeTab extends EventEmitter {
-  private tab: any;
-  private active: boolean;
-
-  constructor(tab: any) {
-    super();
-    this.active = true;
-    this.tab = tab;
-  }
-
-  public async navigate(url: string, opts: navigateOpts = { onload: true }): Promise<void> {
-    log(`navigating to ${url}`);
-    await this.tab.Page.navigate({ url });
-
-    if (opts.onload) {
-      await this.tab.Page.loadEventFired();
-      return;
-    }
-
-    return;
-  }
-
-  public async evaluate(expression: Function, ...args): Promise<any> {
-    const stringifiedArgs = args.map((arg) => JSON.stringify(arg)).join(',');
-    const script = `(${expression.toString()})(${stringifiedArgs})`;
-    
-    log(`executing script: ${script}`);
-    
-    return this.tab.Runtime.evaluate({ expression: script, returnValue: true });
-  }
-
-  private async getSelectorId(selector: string): Promise<number | null> {
-    log(`getting selector '${selector}'`);
-    const document = await this.tab.DOM.getDocument();
-
-    const { nodeId } = await this.tab.DOM.querySelector({
-      nodeId: document.root.nodeId,
-      selector,
-    });
-
-    return nodeId;
-  }
-
-  public async screenShot(filePath: string): Promise<any> {
-    if (!path.isAbsolute(filePath)) {
-      throw new Error(`Filepath is not absolute: ${filePath}`);
-    }
-    log(`capturing screenshot ${filePath}`);
-
-    const base64Image = await this.tab.Page.captureScreenshot();
-    const buffer = new Buffer(base64Image.data, 'base64');
-
-    return fs.writeFileSync(filePath, buffer, { encoding: 'base64' });
-  }
-
-  public async pdf(filePath: string): Promise<any> {
-    if (!path.isAbsolute(filePath)) {
-      throw new Error(`Filepath is not absolute: ${filePath}`);
-    }
-    log(`capturing PDF ${filePath}`);
-
-    const base64Image = await this.tab.Page.printToPDF();
-    const buffer = new Buffer(base64Image.data, 'base64');
-
-    return fs.writeFileSync(filePath, buffer, { encoding: 'base64' });
-  }
-
-  public async setWindowSize(width:number, height:number): Promise<any> {
-    log(`setting window size ${width}x${height}`);
-
-    return this.tab.Emulation.setVisibleSize({ width: width, height: height });
-  }
-
-  public async exists(selector: string): Promise<boolean> {
-    log(`checking if '${selector}' exists`);
-    
-    return !!await this.getSelectorId(selector);
-  }
-
-  public async getHTML(selector: string): Promise<string | null> {
-    log(`getting '${selector}' HTML`);
-
-    const nodeId = await this.getSelectorId(selector);
-
-    if (!nodeId) {
-      return null;
-    }
-
-    const { outerHTML } = await this.tab.DOM.getOuterHTML({ nodeId });
-
-    return outerHTML;
-  }
-
-  public async trigger(eventName: triggerEvents, selector: string): Promise<any> {
-    log(`triggering '${eventName}' on '${selector}'`);
-    let eventClass = '';
-
-    switch (eventName) {
-      case 'click':
-      case 'mousedown':
-      case 'mouseup':
-      case 'mouseover':
-        eventClass = 'MouseEvents';
-        break;
-
-      case 'touchstart':
-      case 'touchend':
-      case 'touchcancel':
-      case 'touchmove':
-        eventClass = 'TouchEvents';
-        break;
-
-      case 'focus':
-      case 'change':
-      case 'blur':
-      case 'select':
-        eventClass = 'HTMLEvents';
-        break;
-
-      default:
-        throw `chrome#trigger: Couldn't handle event ${eventName} on selector ${selector}.`;
-    }
-
-    const expression = function(selector, eventName, eventClass) {
-      const node = document.querySelector(selector);
-      const doc = node && node.ownerDocument ? node.ownerDocument : node;
-      const e = doc && doc.createEvent(eventClass);
-      e.initEvent(eventName, true);
-      e.synthetic = true;
-      node.dispatchEvent(e, true);
-    };
-
-    return this.evaluate(expression, selector, eventName, eventClass);
-  }
-
-  public done(): void {
-    if (this.active) {
-      log(`clearing tab`);
-      this.active = false;
-      this.tab = null;
-      this.emit(events.done);
-    }
-  }
 }
 
 export class Chrome {
@@ -223,13 +54,13 @@ export class Chrome {
       .map((_value, key) => `--${_.kebabCase(key)}`)
       .value();
 
-    log(`launching with args ${chromeFlags.join(' ')}`);
+    log(`launching host with args ${chromeFlags.join(' ')}`);
 
     // Boot Chrome
     const browser = await chromeLauncher.launch({ chromeFlags });
     const cdp = await CDP({ target: `ws://localhost:${browser.port}/devtools/browser` });
 
-    log(`launched on port ${browser.port}`);
+    log(`launched host on port ${browser.port}`);
 
     this.kill = browser.kill;
     this.host = cdp;
@@ -261,16 +92,17 @@ export class Chrome {
       newTab.CSS.enable(),
     ]);
 
-    const tab = new ChromeTab(newTab);
+    const tab = new ChromeTab(newTab, targetId);
 
-    tab.on(events.done, this.onTabClose);
+    tab.on(events.done, this.onTabClose.bind(this));
 
     this.activeTabs++;
     return tab;
   }
 
-  public onTabClose(): void {
-    log(`tab closed`);
+  public onTabClose(targetId: string): void {
+    this.chrome.Target.closeTarget({ targetId });
+    log(`tab ${targetId} closed`);
     this.activeTabs--;
   }
 
@@ -287,11 +119,7 @@ export class Chrome {
   }
 
   public getIsBusy(): boolean {
-    const isBusy:boolean = this.maxActiveTabs === this.activeTabs;
-
-    log(`instance ${isBusy ? 'still has' : 'has no'} capacity`);
-
-    return isBusy;
+    return this.maxActiveTabs === this.activeTabs;
   }
 
   public getIsExpired(): boolean {
